Flash interface tested

+++ Flash_Honda/include/FLITF.h
@@ -14,7 +14,7 @@ typedef unsigned short int uint_16t;
 typedef unsigned long int uint_32t;
 typedef signed char int_8t;
 typedef signed short int int_16t;
-typedef signed short int int_32t;
+typedef signed long int int_32t;
 
 #define OK 		0
 #define NOT_OK 	1
